Force numpy :cop:

+++ pytorch_grad_cam/utils/image.py
@@ -195,9 +195,9 @@ def scale_cam_image(cam, target_size=None, compute_device=None):
         if target_size is not None:
             img = cv2.resize(img.cpu().numpy(), target_size)
         result.append(img)
-    result = result.to(torch.float32) # Do we need float32s?
 
-    return torch.tensor(result).to(compute_device)
+    # Do we need float32s?
+    return torch.tensor(result).to(torch.float32).to(compute_device)
 
 
 def scale_accross_batch_and_channels(tensor, target_size):
